+++ sanic_cors/__init__.py
@@ -5,7 +5,7 @@
     Sanic-CORS is a simple extension to Sanic allowing you to support cross
     origin resource sharing (CORS) using a simple decorator.
 
-    :copyright: (c) 2021 by Ashley Sommer (based on flask-cors by Cory Dolphin).
+    :copyright: (c) 2022 by Ashley Sommer (based on flask-cors by Cory Dolphin).
     :license: MIT, see LICENSE for more details.
 """
 from .decorator import cross_origin