+++ tests/decorator/test_allow_headers.py
@@ -14,7 +14,7 @@ from sanic_cors.core import *
 
 class AllowHeadersTestCaseIntegration(SanicCorsTestCase):
     def setUp(self):
-        self.app = Sanic(__name__)
+        self.app = Sanic(__name__.replace(".","-"))
 
         @self.app.route('/test_default', methods=['GET', 'HEAD', 'OPTIONS'])
         @cross_origin(self.app)