+++ tests/decorator/test_duplicate_headers.py
@@ -25,7 +25,7 @@ from sanic_cors.core import *
 
 class AllowsMultipleHeaderEntries(SanicCorsTestCase):
     def setUp(self):
-        self.app = Sanic(__name__)
+        self.app = Sanic(__name__.replace(".","-"))
 
         @self.app.route('/test_multiple_set_cookie_headers')
         @cross_origin(self.app)