+++ tests/decorator/test_expose_headers.py
@@ -15,7 +15,7 @@ from sanic_cors.core import *
 
 class ExposeHeadersTestCase(SanicCorsTestCase):
     def setUp(self):
-        self.app = Sanic(__name__)
+        self.app = Sanic(__name__.replace(".","-"))
 
         @self.app.route('/test_default', methods=['GET', 'HEAD', 'OPTIONS'])
         @cross_origin(self.app)