+++ tests/decorator/test_vary_header.py
@@ -28,7 +28,7 @@ from sanic_cors import *
 
 class VaryHeaderTestCase(SanicCorsTestCase):
     def setUp(self):
-        self.app = Sanic(__name__)
+        self.app = Sanic(__name__.replace(".","-"))
 
         @self.app.route('/', methods=['GET', 'HEAD', 'OPTIONS'])
         @cross_origin(self.app)